backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
Pillow
pdfkit
